tests/wpt/tests/css/css-cascade/parsing/all-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Cascading and Inheritance Level 3: parsing all with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-cascade-3/#propdef-all">
<meta name="assert" content="all supports only the grammar 'initial | inherit | unset'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("all", "auto");
test_invalid_value("all", "none");
test_invalid_value("all", "filter");
test_invalid_value("all", "unset inherit");
test_invalid_value("all", "inherit initial");
test_invalid_value("all", "initial unset");
test_invalid_value("all", "opacity transform");
</script>
</body>
</html>
